requirements.txt
pdfminer.six
Pillow
pytesseract
PyMuPDF
boto3
python-dotenv
flask
werkzeug
